requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
frozenlist==1.3.3
gevent==22.10.2
greenlet==2.0.1
idna==3.4
multidict==6.0.2
py-cord==2.2.2
pycares==4.2.2
pycodestyle==2.10.0
pycparser==2.21
python-dotenv==0.21.0
pymongo==4.3.3
requests==2.28.1
tinydb==4.7.0
tomli==2.0.1
urllib3==1.26.12
yarl==1.8.1
zope.event==4.5.0
zope.interface==5.5.2
